requirements.txt
pymongo
requests
beautifulsoup4
selenium
pandas
openpyxl
googlemaps
python-dotenv
pillow
